[vexec] Move QueryPlan and methods to its own file
